+++ src/searchHashtag.js
@@ -1,3 +1,5 @@
+import { DOMParser } from "https://deno.land/x/deno_dom/deno-dom-wasm.ts";
+
 /**
  * キーワードから関連ハッシュタグを検索
  *
@@ -5,25 +7,24 @@
  * @param {number} count 取得件数
  * @returns ハッシュタグの配列
  */
-export async function getHashtag(keyword,step){
-    const url_base = "https://instagram.userlocal.jp/hashtags/search?hashtags=";
-    keyword = encodeURIComponent(keyword);
-    const url = url_base+keyword;
-    const res = await fetch(url);
-    const html = await res.text();
-    const par = new DOMParser();
-    const doc = par.parseFromString(html,"text/html");
-    //html化
-    const arr = [];
-    let i = 0;
-    let count = 0;
-    const queryhash = doc.querySelectorAll("span")
-    while(count < step){
-        if(queryhash[i].innerHTML[0] == "#"){
-            arr.push(queryhash[i].innerHTML);
-            count++;
-        }
-        i++;
+export async function getHashtag(keyword) {
+  const url = `https://instagram.userlocal.jp/hashtags/search?hashtags=${encodeURIComponent(
+    keyword
+  )}`;
+  const res = await fetch(url);
+  const html = await res.text();
+  //console.log(html);
+  const par = new DOMParser();
+  const doc = par.parseFromString(html, "text/html");
+  const arr = [];
+  const queryhash = doc.querySelectorAll("span");
+
+  for (const hash of queryhash) {
+    if (!hash?.innerHTML) continue;
+    if (hash.innerHTML[0] === "#") {
+      arr.push(hash.innerHTML.replace(/#/, ""));
     }
-    return arr;
+  }
+
+  return arr;
 }
